explicit cast to int
svn path=/trunk/kdelibs/; revision=196821

+++ kdeui/ktip.cpp
@@ -140,7 +140,7 @@ KTipDialog::KTipDialog(KTipDatabase *db, QWidget *parent, const char *name)
 
     _blendedColor = KGlobalSettings::activeTitleColor();
     _blendedColor.hsv(&h,&s,&v);
-    _blendedColor.setHsv(h,s*(71/76.0),v*(67/93.0));
+    _blendedColor.setHsv(h, int(s*(71/76.0)), int(v*(67/93.0)));
 
     if (!isTipDialog)
     {
@@ -154,7 +154,7 @@ KTipDialog::KTipDialog(KTipDatabase *db, QWidget *parent, const char *name)
 
     _baseColor = KGlobalSettings::alternateBackgroundColor();
     _baseColor.hsv(&h,&s,&v);
-    _baseColor.setHsv(h,s*(10/6.0),v*(93/99.0));
+    _baseColor.setHsv(h, int(s*(10/6.0)), int(v*(93/99.0)));
 
     _textColor = KGlobalSettings::textColor();
